+++ lib/openapi_parser/spec_validator/rules/path_items_in_30.rb
@@ -6,7 +6,7 @@ module OpenAPIParser
       # mismatch.
       class PathItemsIn30 < Rule
         def check(root)
-          return [] unless version == :v3_0
+          return [] unless version_before?('3.1')
 
           components = root.components
           return [] unless components